requirements.txt
Django==1.9.5
Markdown==2.6.6
argparse==1.2.1
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
wsgiref==0.1.2
